enable disable log feature added

+++ index.js
@@ -8,43 +8,45 @@ const utils = require('./utils');
 let date = new Date();
 
 var logstack = {
-    createLog: function (pathDirectory, loggerCount, log) {
+    createLog: function (pathDirectory, loggerCount, log, enable = true) {
 
-        // Appending log message with Time of logging
-        let logMessage = utils.formatTime(date) + " --> " + log + "\n\n";
+        if (enable) {
+            // Appending log message with Time of logging
+            let logMessage = utils.formatTime(date) + " --> " + log + "\n\n";
 
-        // Filename is according to current date
-        let fileName = utils.formatDate(date);
+            // Filename is according to current date
+            let fileName = utils.formatDate(date);
 
-        /**
-         * Reading the Number of files to maintain the file counts in directory
-         */
-        fs.readdir(pathDirectory, (err, files) => {
-            // File path with file name
-            let filePath = pathDirectory + '/' + fileName;
-            if (err) {
-                throw err;
-            } else {
-                /**
-                 * If the count is greater than logger count then
-                 * deleting old files and insert new current date files
-                 * Else appending the log message to same date file
-                 */
-                if (files.length > loggerCount) {
-                    fs.unlink(pathDirectory + '/' + files[0], (err) => {
-                        if (err) throw err;
+            /**
+             * Reading the Number of files to maintain the file counts in directory
+             */
+            fs.readdir(pathDirectory, (err, files) => {
+                // File path with file name
+                let filePath = pathDirectory + '/' + fileName;
+                if (err) {
+                    throw err;
+                } else {
+                    /**
+                     * If the count is greater than logger count then
+                     * deleting old files and insert new current date files
+                     * Else appending the log message to same date file
+                     */
+                    if (files.length > loggerCount) {
+                        fs.unlink(pathDirectory + '/' + files[0], (err) => {
+                            if (err) throw err;
 
+                            fs.appendFile(filePath, logMessage, (err) => {
+                                if (err) throw err;
+                            });
+                        });
+                    } else {
                         fs.appendFile(filePath, logMessage, (err) => {
                             if (err) throw err;
                         });
-                    });
-                } else {
-                    fs.appendFile(filePath, logMessage, (err) => {
-                        if (err) throw err;
-                    });
+                    }
                 }
-            }
-        });
+            });
+        }
     }
 }
